fix(payments): montos de caja como enteros, eliminar floats (P1-005)
CIERRE DE BLOQUEADOR P1-005:
Los FormRequests de apertura/cierre de caja aceptaban montos decimales
(numeric) que luego se procesaban como floats, violando el contrato
ADR-018 "Money 100% INTEGER" y causando truncamientos en difference.

CAMBIOS EN FORMREQUESTS:
✅ OpenCashSessionRequest.php:
   - Antes: 'opening_amount' => ['required', 'numeric', 'min:0']
   - Después: 'opening_amount' => ['required', 'integer', 'min:0']

✅ CloseCashSessionRequest.php:
   - Antes: 'closing_amount' => ['required', 'numeric', 'min:0']
   - Después: 'closing_amount' => ['required', 'integer', 'min:0']

CAMBIOS EN SERVICIO DE DOMINIO:
✅ CashSessionService.php:
   - openSession(): float $openingAmount → int $openingAmount
   - closeSession(): float $closingAmount → int $closingAmount
   - Cálculo de diferencia: eliminado round($closing - $expected, 2)
   - Nuevo cálculo: $difference = $closingAmount - $expected; (siempre int)

TEST DE VALIDACIÓN (9 tests, 22 assertions):
✅ Análisis estático: verifica reglas 'integer' en ambos requests
✅ Comportamiento: valida que 10000.50 sea rechazado con 422
✅ Comportamiento: valida que 10000 sea aceptado
✅ Lógica de dominio: verifica que difference se calcule sin round()
✅ Lógica de dominio: verifica que la resta de enteros siempre sea int

IMPACTO:
✅ API rechaza explícitamente montos de caja decimales
✅ Cálculo de diferencias de arqueo es 100% preciso (sin pérdida de decimales)
✅ Contrato ADR-018 garantizado en flujo de caja

VERIFICACIÓN:
✅ 9/9 tests de P1-005 pasando
✅ Suite completa sin regresiones

Refs: ADR-018 Money and Tax Architecture

+++ app/Modules/Payments/Interfaces/Requests/CloseCashSessionRequest.php
@@ -14,7 +14,7 @@ class CloseCashSessionRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'closing_amount' => ['required', 'numeric', 'min:0'],
+            'closing_amount' => ['required', 'integer', 'min:0'],
             'notes' => ['nullable', 'string', 'max:500'],
         ];
     }
